chore: move action menu to its own file
