+++ src/navigation.ts
@@ -1259,113 +1259,6 @@ export class Navigation {
   }
 
   /**
-   * Show the action menu for the current node.
-   *
-   * The action menu will contain entries for relevant actions for the
-   * node's location.  If the location is a block, this will include
-   * the contents of the block's context menu (if any).
-   *
-   * Returns true if it is possible to open the action menu in the
-   * current location, even if the menu was not opened due there being
-   * no applicable menu items.
-   */
-  openActionMenu(workspace: Blockly.WorkspaceSvg): boolean {
-    let menuOptions: Array<
-      | Blockly.ContextMenuRegistry.ContextMenuOption
-      | Blockly.ContextMenuRegistry.LegacyContextMenuOption
-    > = [];
-    let rtl: boolean;
-
-    const cursor = workspace.getCursor();
-    if (!cursor) throw new Error('workspace has no cursor');
-    const node = cursor.getCurNode();
-    const nodeType = node.getType();
-    switch (nodeType) {
-      case Blockly.ASTNode.types.BLOCK:
-        const block = node.getLocation() as Blockly.BlockSvg;
-        rtl = block.RTL;
-        // Reimplement BlockSvg.prototype.generateContextMenu as that
-        // method is protected.
-        if (!workspace.options.readOnly && block.contextMenu) {
-          menuOptions =
-            Blockly.ContextMenuRegistry.registry.getContextMenuOptions(
-              Blockly.ContextMenuRegistry.ScopeType.BLOCK,
-              {block},
-            );
-
-          // Allow the block to add or modify menuOptions.
-          block.customContextMenu?.(menuOptions);
-        }
-        // End reimplement.
-        break;
-
-      // case Blockly.ASTNode.types.INPUT:
-      case Blockly.ASTNode.types.NEXT:
-      case Blockly.ASTNode.types.PREVIOUS:
-      case Blockly.ASTNode.types.INPUT:
-        const connection = node.getLocation() as Blockly.Connection;
-        rtl = connection.getSourceBlock().RTL;
-
-        // Slightly hacky: get insert action from registry.  Hacky
-        // because registry typings don't include {connection: ...} as
-        // a possible kind of scope.
-        const insertAction =
-          Blockly.ContextMenuRegistry.registry.getItem('insert');
-        if (!insertAction) throw new Error("can't find insert action");
-
-        const pasteAction = Blockly.ContextMenuRegistry.registry.getItem(
-          'blockPasteFromContextMenu',
-        );
-        if (!pasteAction) throw new Error("can't find paste action");
-        const possibleOptions = [insertAction, pasteAction /* etc.*/];
-
-        // Check preconditions and get menu texts.
-        const scope = {
-          connection,
-        } as unknown as Blockly.ContextMenuRegistry.Scope;
-        for (const option of possibleOptions) {
-          const precondition = option.preconditionFn(scope);
-          if (precondition === 'hidden') continue;
-          const displayText =
-            typeof option.displayText === 'function'
-              ? option.displayText(scope)
-              : option.displayText;
-          menuOptions.push({
-            text: displayText,
-            enabled: precondition === 'enabled',
-            callback: option.callback,
-            scope,
-            weight: option.weight,
-          });
-        }
-        break;
-
-      default:
-        console.info(`No action menu for ASTNode of type ${nodeType}`);
-        return false;
-    }
-
-    if (!menuOptions?.length) return true;
-    const fakeEvent = fakeEventForNode(node);
-    Blockly.ContextMenu.show(fakeEvent, menuOptions, rtl, workspace);
-    setTimeout(() => {
-      Blockly.WidgetDiv.getDiv()
-        ?.querySelector('.blocklyMenu')
-        ?.dispatchEvent(
-          new KeyboardEvent('keydown', {
-            key: 'ArrowDown',
-            code: 'ArrowDown',
-            keyCode: Blockly.utils.KeyCodes.DOWN,
-            which: Blockly.utils.KeyCodes.DOWN,
-            bubbles: true,
-            cancelable: true,
-          }),
-        );
-    }, 10);
-    return true;
-  }
-
-  /**
    * Pastes the copied block to the marked location if possible or
    * onto the workspace otherwise.
    *
@@ -1425,108 +1318,6 @@ export class Navigation {
       this.removeWorkspace(workspace);
     }
   }
-}
-
-/**
- * Create a fake PointerEvent for opening the action menu for the
- * given ASTNode.
- *
- * @param node The node to open the action menu for.
- * @returns A synthetic pointerdown PointerEvent.
- */
-function fakeEventForNode(node: Blockly.ASTNode): PointerEvent {
-  switch (node.getType()) {
-    case Blockly.ASTNode.types.BLOCK:
-      return fakeEventForBlockNode(node);
-    case Blockly.ASTNode.types.NEXT:
-    case Blockly.ASTNode.types.PREVIOUS:
-    case Blockly.ASTNode.types.INPUT:
-      return fakeEventForConnectionNode(node);
-    default:
-      throw new TypeError('unhandled node type');
-  }
-}
-
-/**
- * Create a fake PointerEvent for opening the action menu for the
- * given ASTNode of type BLOCK.
- *
- * @param node The node to open the action menu for.
- * @returns A synthetic pointerdown PointerEvent.
- */
-function fakeEventForBlockNode(node: Blockly.ASTNode): PointerEvent {
-  if (node.getType() !== Blockly.ASTNode.types.BLOCK) {
-    throw new TypeError('can only create PointerEvents for BLOCK nodes');
-  }
-
-  // Get the location of the top-left corner of the block in
-  // screen coordinates.
-  const block = node.getLocation() as Blockly.BlockSvg;
-  const blockCoords = Blockly.utils.svgMath.wsToScreenCoordinates(
-    block.workspace,
-    block.getRelativeToSurfaceXY(),
-  );
-
-  // Prefer a y position below the first field in the block.
-  const fieldBoundingClientRect = block.inputList
-    .filter((input) => input.isVisible())
-    .flatMap((input) => input.fieldRow)
-    .filter((f) => f.isVisible())[0]
-    ?.getSvgRoot()
-    ?.getBoundingClientRect();
-
-  const clientY =
-    fieldBoundingClientRect && fieldBoundingClientRect.height
-      ? fieldBoundingClientRect.y + fieldBoundingClientRect.height
-      : blockCoords.y + block.height;
-
-  // Create a fake event for the action menu code to work from.
-  return new PointerEvent('pointerdown', {
-    clientX: blockCoords.x + 5,
-    clientY: clientY + 5,
-  });
-}
-
-/**
- * Create a fake PointerEvent for opening the action menu for the
- * given ASTNode of type NEXT, PREVIOUS or INPUT.
- *
- * For now this just puts the action menu in the same place as the
- * context menu for the source block.
- *
- * @param node The node to open the action menu for.
- * @returns A synthetic pointerdown PointerEvent.
- */
-function fakeEventForConnectionNode(node: Blockly.ASTNode): PointerEvent {
-  if (
-    node.getType() !== Blockly.ASTNode.types.NEXT &&
-    node.getType() !== Blockly.ASTNode.types.PREVIOUS &&
-    node.getType() !== Blockly.ASTNode.types.INPUT
-  ) {
-    throw new TypeError('can only create PointerEvents for connection nodes');
-  }
-
-  const connection = node.getLocation() as Blockly.Connection;
-  const block = connection.getSourceBlock();
-  const workspace = block.workspace as Blockly.WorkspaceSvg;
-
-  if (typeof connection.x !== 'number') {
-    // No coordinates for connection?  Fall back to the parent block.
-    const blockNode = new Blockly.ASTNode(Blockly.ASTNode.types.BLOCK, block);
-    return fakeEventForBlockNode(blockNode);
-  }
-  const connectionWSCoords = new Blockly.utils.Coordinate(
-    connection.x,
-    connection.y,
-  );
-  const connectionScreenCoords = Blockly.utils.svgMath.wsToScreenCoordinates(
-    workspace,
-    connectionWSCoords,
-  );
-  return new PointerEvent('pointerdown', {
-    clientX: connectionScreenCoords.x + 5,
-    clientY: connectionScreenCoords.y + 5,
-  });
 }
 
 /**
